refactor(services): Remove legacy services and update references
