+++ src/pages/login.tsx
@@ -2,9 +2,11 @@ import { useState } from "react";
 import { Button, Input, Text, VStack, Center, Heading, Field, Image } from "@chakra-ui/react";
 import { PasswordInput } from "@/components/ui/password-input";
 import { useNavigate } from "react-router";
-import { login } from "@/api/user/user-service";
 import { storage } from "@/utils/storage/user-storage";
 import { getUserIdFromAccessToken } from "@/utils/jwt";
+import { login } from "@/api/user/user-access-service";
+import { create } from "@bufbuild/protobuf";
+import { TokenPairSchema } from "@/proto-generated/nori/v0/user/access/token_pairs_pb";
 
 const LoginPage = () => {
   const [flag, setFlag] = useState(1);
@@ -98,7 +100,11 @@ const Login = () => {
     }
 
     try {
-      const tokenPair = await login(account, password);
+      const userTokenPair = await login(account, password);
+      const tokenPair = create(TokenPairSchema, {
+        refreshToken: userTokenPair.refreshToken,
+        accessToken: userTokenPair.accessToken
+      })
       console.log("get token pair: ", tokenPair);
       const userid = getUserIdFromAccessToken(tokenPair.accessToken); 
       console.log("get userid: ", userid);
