+++ src/utils/storage/user-storage.ts
@@ -1,5 +1,6 @@
-import { TokenPair } from "@/proto-generated/nori/v0/user/token_pair_pb";
+import { TokenPair, TokenPairSchema, UserTokenPair } from "@/proto-generated/nori/v0/user/access/token_pairs_pb";
 import { getUserIdFromAccessToken } from "../jwt";
+import { create } from "@bufbuild/protobuf";
 
 interface UserAuth {
   userId: bigint;
@@ -28,14 +29,35 @@ export const storage = {
     localStorage.removeItem("userAuth");
   },
 
-  saveToken: (inputTokenPair: TokenPair) => {
+  saveToken: (inputToken: TokenPair | UserTokenPair) => {
     try {
-      const userId = getUserIdFromAccessToken(inputTokenPair.accessToken);
-      storage.setUserAuth({ userId: userId.id, tokenPair: inputTokenPair });
+      if ('userId' in inputToken) {
+        // 處理 UserTokenPair
+        const userId = inputToken.userId;
+        if (!userId) throw new Error("userId is null or undifinded");
+        const tokenPair = create(TokenPairSchema, {
+          accessToken: inputToken.accessToken,
+          refreshToken: inputToken.refreshToken
+        })
+        storage.setUserAuth({
+          userId: userId.id,
+          tokenPair: tokenPair
+        });
+      } else {
+        // 處理 TokenPair
+        const userId = getUserIdFromAccessToken(inputToken.accessToken);
+        const tokenPair = create(TokenPairSchema, {
+          accessToken: inputToken.accessToken,
+          refreshToken: inputToken.refreshToken
+          
+        })
+        storage.setUserAuth({
+          userId: userId.id,
+          tokenPair: tokenPair
+        });
+      }
     } catch (error) {
       console.error("save token error: ", error);
     }
-    
-
   }
 };
